requirements.txt
streamlit
openai
langchain
chromadb
python-dotenv